Improved the terrain shader
Added tiling, smoothness and normal functionality for layer 1 of the terrain shader. As well as albedo for Layer 2.

If these extra functionalities are needed for layer 2 they can be easily added, but are only present on layer 1 to see if anything needs to be improved, and to keep the shader graph from becoming overly complex.